refactor: 비즈니스 프로필페이지 파일구조 변경 (#252)
